Split generic custom-video-source out of vision-camera-source
The vision-camera dep was used in only 3 files; the rest (the custom-video
track lifecycle hooks + the WebGPU camera toolkit) is source-agnostic.
Extract it into a new package @fishjam-cloud/react-native-custom-video-source
that publishes your own frames to Fishjam from any source, and make the
VisionCamera package a thin adapter on top of it.

- new package: internal/{useManagedForwardTrack,useManagedPooledTrack,toError}
  + webgpu/* (camera shader bindings, passthrough, crop, render context,
  required features, device, texture resolver, runtime). Owns typegpu +
  unplugin-typegpu (the TGSL shaders live here now).
- vision-camera-source keeps only the VC-coupled files (useVisionCameraSource,
  useVisionCameraWebGpuSource, frameTimestamp, orientation), depends on the
  generic package (workspace:*), and its /webgpu re-exports the generic
  toolkit. Dropped its now-unused typegpu/unplugin-typegpu deps + babel plugin.
- register the new workspace in the root package.json.

Both packages typecheck, build (bob), and lint clean.
