+++ package.json
@@ -9,6 +9,7 @@
     "packages/react-client",
     "packages/webrtc-client",
     "packages/mobile-client",
+    "packages/react-native-custom-video-source",
     "packages/react-native-vision-camera-source",
     "packages/react-native-webrtc",
     "examples/react-client/*",
