+++ packages/react-native-custom-video-source/.eslintignore
@@ -1,0 +1,3 @@
+dist
+build
+coverage
