+++ packages/react-native-custom-video-source/.prettierrc
@@ -1,0 +1,9 @@
+{
+  "bracketSameLine": true,
+  "printWidth": 120,
+  "quoteProps": "consistent",
+  "singleQuote": true,
+  "tabWidth": 2,
+  "trailingComma": "all",
+  "useTabs": false
+}
