+++ packages/react-native-custom-video-source/src/internal/toError.ts
