+++ packages/react-native-custom-video-source/src/webgpu/cameraShaderBindings.ts
