+++ packages/react-native-custom-video-source/typedoc.json
@@ -1,0 +1,6 @@
+{
+  "$schema": "https://typedoc.org/schema.json",
+  "includeVersion": true,
+  "entryPoints": ["src/index.ts", "src/webgpu.ts"],
+  "readme": "none"
+}
